updated course catalog to fetch data from google sheet
